+++ src/components/Header/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'gatsby';
 
-import './header.css';
+import './Header.css';
 
 const Header = ({ siteData }) => (
   <header
