+++ src/components/Layout/Layout.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { StaticQuery, graphql } from 'gatsby';
 
-import Header from './header';
-import './layout.css';
+import Header from '../Header';
+import './Layout.css';
 
 const Layout = ({ children, data }) => (
   <StaticQuery
