+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { graphql } from 'gatsby';
 
-import Layout from '../components/layout';
+import Layout from '../components/Layout';
 
 const IndexPage = ({ data }) => (
   <Layout>
